[fix] 인터페이스 분리 api #13
- OrderService 주석 추가
